Refactored png levels into their own class
Simplifies a lot of the level loading logic, as well as making Level a
lot more abstract. This prepares for making it properly abstract later
on, which will be a prerequisite for making levels with a full color
palette. Since this seems like a nice side effect of using png files, I
intend to pursue this path.